+++ UC_StrParser.py
@@ -1,0 +1,18 @@
+from UC_Errors import *
+
+def parseUnitStr(toParse):
+	unitMap = {}
+	toParse = toParse.strip()
+	components = toParse.split(' ')
+	for component in components:
+		vals = component.split('^')
+		sym = vals[0]
+		if not (sym in unitMap): unitMap[sym] = 0
+		if len(vals) == 1:
+			unitMap[sym] += 1
+		elif len(vals) == 2:
+			unitMap[sym] += int(vals[1])
+		else:
+			raise UnitError(f"Invalid unit: '{component}'")
+	
+	return unitMap
