+++ UC_Utils.py
@@ -12,21 +12,24 @@ def stripPrefix(units, string):
 	return string[0:len(string)-len(longestSuffix)], longestSuffix
 
 # Topological sort implemented using DFS
-def topologicalSortVisit(units, derivedUnit: str, toSort: list, sortedValues: list, visited: set):
+def topologicalSortVisit(units: dict, derivedUnit: str, toSort: list, sortedValues: list, visited: set):
 	prefix, derivedUnit = stripPrefix(units, derivedUnit)
 	if derivedUnit in visited: return
 	visited[derivedUnit] = True
 
 	if derivedUnit in units:
-		for baseUnit in self.units[derivedUnit].baseUnits.keys():
-			self.topologicalSortVisit(baseUnit, toSort, sortedValues, visited)
+		for baseUnit in units[derivedUnit].baseUnits.keys():
+			topologicalSortVisit(units, baseUnit, toSort, sortedValues, visited)
 
 	if derivedUnit in toSort: sortedValues.insert(0, derivedUnit)
 
 # Perform a topological sort over the conversions
-def topologicalSort(toSort: list):
+def topologicalSort(units: dict, toSort: list):
 	sortedValues = []
 	visited = {}
 	for derivedUnit in toSort:
-		topologicalSortVisit(derivedUnit, toSort, sortedValues, visited)
+		topologicalSortVisit(units, derivedUnit, toSort, sortedValues, visited)
 	return sortedValues
+
+# TODO: Implement this!
+def validate(units, conversions, prefixes): pass
